% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-north.R
\name{add_north}
\alias{add_north}
\title{Add a north symbol to a ggplot object, with automated detection of axis limits}
\usage{
add_north(p, x = NULL, y = NULL, ...)
}
\arguments{
\item{p}{ggplot object}

\item{x}{Argument passed to north2: Number between 0 and 1 to indicate the x axis position of the symbol's bottom left corner. 0 is the left side and 1 the right side. Defaults to \code{NULL}, which means \code{\link{north}} will be used.}

\item{y}{Argument passed to north2: Number between 0 and 1 to indicate the y axis position of the symbol's bottom left corner. 0 is the bottom and 1 the top. Defaults to \code{NULL}, which means \code{\link{north}} will be used.}

\item{...}{arguments passed to \code{\link{north}} or \code{\link{north2}}}
}
\value{
ggplot object with the north symbol
}
\description{
This function uses the ggplot object as an argument, enabling automated detection of the axis limits, and making it pipeable.
}
\examples{
library(rgdal); library(broom); library(ggplot2); library(magrittr)
dsn <- system.file('extdata', package = 'ggsn')

## Map in geographic coordinates.
map <- readOGR(dsn, 'sp')
map@data$id <- 1:nrow(map@data)
map.df <- merge(tidy(map), map, by = 'id')
 
{ggplot(data = map.df, aes(long, lat, group = group, fill = nots)) +
   geom_polygon() +
   coord_equal() +
   geom_path() +
   scale_fill_brewer(name = 'Animal abuse\\nnotifications', palette = 8)} \%>\% 
   add_north()
}

