% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-scalebar.R
\name{add_scalebar}
\alias{add_scalebar}
\title{Add a scale bar to a ggplot object, with automated detection of axis limits}
\usage{
add_scalebar(p, ...)
}
\arguments{
\item{p}{ggplot object}

\item{...}{arguments passed to \code{\link{scalebar}}}
}
\value{
ggplot object with the scalebar
}
\description{
This function uses the ggplot object as an argument, enabling automated detection of the axis limits, and making it pipeable.
}
\examples{
library(rgdal); library(broom); library(ggplot2); library(magrittr)
dsn <- system.file('extdata', package = 'ggsn')

## Map in geographic coordinates.
map <- readOGR(dsn, 'sp')
map@data$id <- 1:nrow(map@data)
map.df <- merge(tidy(map), map, by = 'id')
 
{ggplot(data = map.df, aes(long, lat, group = group, fill = nots)) +
   geom_polygon() +
   coord_equal() +
   geom_path() +
   scale_fill_brewer(name = 'Animal abuse\\nnotifications', palette = 8)} \%>\% 
   add_scalebar(dist = 5, dd2km = TRUE, model = 'WGS84')
}

