% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank.R
\name{blank}
\alias{blank}
\title{Blank theme}
\usage{
blank()
}
\description{
ggplot blank theme.
}
\examples{
library(sf)
dsn <- system.file('extdata', package = 'ggsn')
map <- st_read(dsn, 'sp')

# If "map" is a "sp" object, convert it to "sf" with map <- st_as_sf(map).

# Map in geographic coordinates
ggplot(map, aes(fill = nots)) +
    geom_sf() +
    scalebar(map, , dist = 5, dd2km = TRUE, model = 'WGS84') +
    blank() +
    scale_fill_brewer(name = 'Animal abuse\\nnotifications', palette = 8)
    
}
