% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggsn-package.R
\docType{package}
\name{ggsn-package}
\alias{ggsn-package}
\title{The ggsn Package}
\description{
North symbols and scale bars for ggplot and ggmap maps. Adds north symbols (18 options) and scale bars in kilometers to maps in geographic or metric coordinates created with ggplot or ggmap.
}
\details{
\tabular{ll}{
Package: \tab ggsn\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2015-08-21\cr
Depends: \tab R (>= 3.2.1)\cr
Imports: \tab ggplot2, png, grid, maptools\cr
Suggests: \tab rgdal, ggmap\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
URL: \tab \url{http://oswaldosantos.github.io/capm}\cr
Author: \tab Oswaldo Santos Baquero \email{oswaldosant@gmail.com}\cr
Maintainer: \tab Oswaldo Santos Baquero \email{oswaldosant@gmail.com}\cr
}
}
\keyword{package}

