% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/northSymbols.R
\name{northSymbols}
\alias{northSymbols}
\title{Available north symbols.}
\usage{
northSymbols()
}
\description{
Displays avaliable north symbols.
}
\note{
The symbols were obtained from QGIS 2.8.1 - Wien.
}
\examples{
northSymbols()
}
\references{
http://www.qgis.org/en/site
}
