% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/symbols.R
\name{symbols}
\alias{symbols}
\title{Available north symbols.}
\usage{
symbols()
}
\description{
Displays avaliable north symbols.
}
\note{
The symbols were obtained from QGIS 2.8.1 - Wien.
}
\examples{
symbols()
}
\references{
http://www.qgis.org/en/site
}

