% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/blank.R
\name{blank}
\alias{blank}
\title{Blank theme}
\usage{
blank()
}
\description{
ggplot blank theme.
}
\examples{
library(rgdal)
dsn <- system.file('extdata', package = 'ggsn')
map <- readOGR(dsn, 'sp')
map@data$id <- 1:nrow(map@data)
map.ff <- fortify(map, region = 'id')
map.df <- merge(map.ff, map@data, by = 'id')

ggplot(map.df, aes(long, lat, group = group, fill = nots)) +
      geom_polygon() +
      coord_equal() +
      geom_path() +
      north(map.df) +
      blank()

ggplot(map.df, aes(long, lat, group = group, fill = nots)) +
      geom_polygon() +
      coord_equal() +
      geom_path() +
      north(map.df)
}

