% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank.R
\name{blank}
\alias{blank}
\title{Blank theme}
\usage{
blank()
}
\description{
ggplot blank theme.
}
\examples{
\dontrun{
library(sf)
data(domestic_violence)
ggplot(domestic_violence, aes(fill = Scaled)) +
    geom_sf() +
    blank() +
    scale_fill_continuous(low = "#fff7ec", high = "#7F0000")
}
}
