% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsn-package.R
\docType{package}
\name{ggsn-package}
\alias{ggsn-package}
\title{The ggsn Package}
\description{
North symbols and scale bars for maps created with 'ggplot' or 'ggmap'. Adds north symbols (18 options) and scale bars in kilometers to maps in geographic or metric coordinates created with 'ggplot' or 'ggmap'.
}
\details{
\tabular{ll}{
Package: \tab ggsn\cr
Type: \tab Package\cr
Version: \tab 0.5.0\cr
Date: \tab 2016-11-12\cr
Depends: \tab R (>= 3.4), ggplot2\cr
Imports: \tab grid, ggmap, sf, png, maptools\cr
Suggests: \tab rgdal, broom\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
URL: \tab \url{https://github.com/oswaldosantos/ggsn}\cr
Author: \tab Oswaldo Santos Baquero \email{baquero@usp.br}\cr
Maintainer: \tab Oswaldo Santos Baquero \email{baquero@usp.br}\cr
}
}
\keyword{package}
